@@ -4,8 +4,8 @@ version: 2.0.4+1
 homepage: https://github.com/vikram25897/flutter_cached_video_player
 
 environment:
-  sdk: ">=2.17.0 <4.0.0"
-  flutter: ">=3.0.0"
+  sdk: ">=3.5.0 <4.0.0"
+  flutter: ">=3.24.0"
 
 flutter:
   plugin:
@@ -22,7 +22,7 @@ dependencies:
   flutter:
     sdk: flutter
 
-  video_player_platform_interface: ^5.1.4
+  video_player_platform_interface: ^6.2.3
 
   # The design on https://flutter.dev/go/federated-plugins was to leave
   # this constraint as "any". We cannot do it right now as it fails pub publish
@@ -30,12 +30,12 @@ dependencies:
   # the constraints on the interface pins it.
   # TODO(amirh): Revisit this (either update this part in the design or the pub tool).
   # https://github.com/flutter/flutter/issues/46264
-  video_player_web: ^2.0.10
+  video_player: ^2.9.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.1
+  flutter_lints: ^5.0.0
   pedantic: ^1.11.1
-  pigeon: ^9.2.4
+  pigeon: ^22.5.0
